feat(dungeon): introduce LocalPosition / AbsolutePosition / Module
Phase 1 of the coordinate-types refactor (rpg-api issue #471).

Defines two distinct cube-coordinate types in the dungeon component —
LocalPosition (room-relative) and AbsolutePosition (dungeon-relative) —
so the compiler refuses to substitute one for the other. Module is the
single canonical bridge between the two spaces; LocalToAbsolute and
AbsoluteToLocal are the only legal way to cross the boundary, replacing
the ad-hoc element-wise math (and the cubeY := -cubeX - cubeZ vestiges)
that produced the Round 2 fix train of PRs #459–#468.

Module follows the toolkit pattern: state is loaded from Data and
serialized back via ToData; the data orchestrator owns persistence.

This commit is additive — no existing call sites are migrated yet
(those land in Phases 2 and 3 of the same PR). The only edits to
existing code are int32 boundary casts forced by the new proto Position
shape (rpg-api-protos PR #147 made apiv1alpha1.Position int32 X/Y/Z).
Those casts are clearly marked at the proto/orchestrator/handler seams
and are removed in Phase 3 once entities.Position and dungeon.Position
are deleted.

Open question from plan.md (toolkit type reuse): defined LocalPosition
and AbsolutePosition as independent types rather than aliasing
spatial.CubeCoordinate. Embedding wouldn't strengthen the local-vs-
absolute type distinction (Add/Distance methods would still operate on
the embedded primitive), and keeping them local preserves graduation
flexibility when the dungeon component eventually moves to rpg-toolkit
(issue #479). Documented in coords.go.

Tests:
- coords_test.go covers constructor invariant enforcement and accessor
  sanity (10 cases)
- module_test.go covers load/save round-trip, transform correctness,
  defensive copy on ToData, missing-room error, and nil-receiver
  behavior (13 cases)

Refs: rpg-project/ideas/coordinate-types/plan.md, design.md
Refs #471

+++ internal/entities/encounter_state_builder_test.go
@@ -286,9 +286,9 @@ func (s *EncounterStateBuilderTestSuite) TestBuildEncounterStateData_DoorPositio
 	door := output.EncounterStateData.Doors["conn-1"]
 	s.Require().NotNil(door)
 	s.Require().NotNil(door.Position)
-	s.Equal(float64(2), door.Position.X)
-	s.Equal(float64(-1), door.Position.Y)
-	s.Equal(float64(-1), door.Position.Z)
+	s.Equal(int32(2), door.Position.X)
+	s.Equal(int32(-1), door.Position.Y)
+	s.Equal(int32(-1), door.Position.Z)
 }
 
 func (s *EncounterStateBuilderTestSuite) TestBuildEncounterStateData_DoorNilPosition() {
@@ -365,7 +365,7 @@ func (s *EncounterStateBuilderTestSuite) TestBuildEncounterStateData_RoomsMapPas
 	s.Equal(int32(11), room.Height)
 	s.Equal(apiv1alpha1.GridType_GRID_TYPE_HEX_POINTY, room.GridType)
 	s.Require().NotNil(room.Origin)
-	s.Equal(float64(0), room.Origin.X)
+	s.Equal(int32(0), room.Origin.X)
 	s.Require().Len(room.Walls, 1)
 	s.Equal("stone", room.Walls[0].Material)
 	s.True(room.Walls[0].BlocksMovement)
